+++ iOS-FakeNFT-Extended/Catalog/Collection/CollectionViewModel.swift
